Align GraphQL implementation with specification
- Fix ExecutionDurationInSeconds type: double? -> int? (spec says int)
- Remove Epic entity (not in GraphQL spec data model)
- Fix LargeLanguageModel relationship: 1:1 PK-as-FK -> proper 1:N with ProjectId FK
- Add CreateLargeLanguageModelInput.ProjectId field and update handlers
- Add scoped queries: GetDeliverablesByProjectId, GetAgentTasksByDeliverableId, GetLargeLanguageModelsByProjectId
- Add LargeLanguageModelType.ProjectId GraphQL field
- Create EF migration for schema changes

+++ src/Server/DevStack.Domain/Entities/AgentTask.cs
@@ -35,7 +35,7 @@ public class AgentTask
 
     public int? CompletionTokens { get; set; }
 
-    public double? ExecutionDurationInSeconds { get; set; }
+    public int? ExecutionDurationInSeconds { get; set; }
 
     public string? Model { get; set; }
